# Improve the *local* image caching in `PartialEvaluator.getOperatorList`