@Snuffleupagus 2020-05-28
> Looks good; thanks a lot!

Thanks for landing this; hopefully the recent image-caching improvements will help reduce both rendering times *and* memory usage in lots of PDF files :-)